refactor: rename css files
