+++ src/widgets/header/index.tsx
@@ -1,5 +1,5 @@
 import { ReactNode } from 'react';
-import styles from './index.module.css';
+import styles from './header.module.css';
 
 type HeaderProps = { children: ReactNode };
